requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6 
pytest
pytest-asyncio
openai
httpx
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0
asyncpg==0.29.0
greenlet==3.0.1
GitPython==3.1.40